statistics

+++ src/main/java/shop/dziupla/spring/login/payload/response/StatisticsDTO.java
@@ -1,70 +1,70 @@
 package shop.dziupla.spring.login.payload.response;
 
 public class StatisticsDTO {
-    private int carCount;
-    private int rentedCars;
-    private int servicedCars;
-    private int availableCars;
-    private int clientCount;
-    private int officeCount;
-    private int employeeCount;
+    private long carCount;
+    private long rentedCars;
+    private long servicedCars;
+    private long availableCars;
+    private long clientCount;
+    private long officeCount;
+    private long employeeCount;
     public StatisticsDTO(){
 
     }
 
-    public int getCarCount() {
+    public long getCarCount() {
         return carCount;
     }
 
-    public void setCarCount(int carCount) {
+    public void setCarCount(long carCount) {
         this.carCount = carCount;
     }
 
-    public int getRentedCars() {
+    public long getRentedCars() {
         return rentedCars;
     }
 
-    public void setRentedCars(int rentedCars) {
+    public void setRentedCars(long rentedCars) {
         this.rentedCars = rentedCars;
     }
 
-    public int getServicedCars() {
+    public long getServicedCars() {
         return servicedCars;
     }
 
-    public void setServicedCars(int servicedCars) {
+    public void setServicedCars(long servicedCars) {
         this.servicedCars = servicedCars;
     }
 
-    public int getAvailableCars() {
+    public long getAvailableCars() {
         return availableCars;
     }
 
-    public void setAvailableCars(int availableCars) {
+    public void setAvailableCars(long availableCars) {
         this.availableCars = availableCars;
     }
 
-    public int getClientCount() {
+    public long getClientCount() {
         return clientCount;
     }
 
-    public void setClientCount(int clientCount) {
+    public void setClientCount(long clientCount) {
         this.clientCount = clientCount;
     }
 
-    public int getOfficeCount() {
+    public long getOfficeCount() {
         return officeCount;
     }
 
-    public void setOfficeCount(int officeCount) {
+    public void setOfficeCount(long officeCount) {
         this.officeCount = officeCount;
     }
 
-    public int getEmployeeCount() {
+    public long getEmployeeCount() {
         return employeeCount;
     }
 
-    public void setEmployeeCount(int employeeCount) {
+    public void setEmployeeCount(long employeeCount) {
         this.employeeCount = employeeCount;
     }
 }
